refactor(frontend): move single-route components to route-private _components/
ThemeSwitcher, TitleFilterSidebar and ManufacturerFilterSidebar are each
used by only one route family. Move them (with their fixture/test
companions) under routes/.../_components/ and update the page importers.
