+++ frontend/src/routes/manufacturers/_components/ManufacturerFilterSidebar.dom.test.ts
