+++ frontend/src/routes/manufacturers/_components/ManufacturerFilterSidebar.fixture.svelte
