+++ frontend/src/routes/manufacturers/_components/ManufacturerFilterSidebar.svelte
