+++ frontend/src/routes/style-lab/_components/ThemeSwitcher.svelte
