+++ frontend/src/routes/titles/+page.svelte
@@ -2,7 +2,7 @@
   import client from '$lib/api/client';
   import { unwrapPage } from '$lib/paginated-loader.svelte';
   import FacetedCatalogListing from '$lib/components/pages/listing/FacetedCatalogListing.svelte';
-  import TitleFilterSidebar from '$lib/components/TitleFilterSidebar.svelte';
+  import TitleFilterSidebar from './_components/TitleFilterSidebar.svelte';
   import TitleCard from '$lib/components/collections/cards/TitleCard.svelte';
   import { filtersFromParams, filtersToParams } from '$lib/facet-engine';
   import { titleFilterChips } from './titles-filter-chips';
