+++ frontend/src/routes/titles/_components/TitleFilterSidebar.fixture.svelte
