+++ frontend/src/routes/titles/_components/TitleFilterSidebar.svelte
